Move Spark integration tests into Spark projects
This change moves the Spark int-tests into the Spark main module, because the reason for the separate project, the Quarkus build, no longer exists.
