+++ bom/build.gradle.kts
@@ -104,13 +104,10 @@ dependencies {
     api(project(":polaris-hms-testcontainer"))
 
     api(project(":polaris-spark-3.5_2.12"))
-    api(project(":polaris-spark-integration-3.5_2.12"))
     val ideaActive = providers.systemProperty("idea.active").getOrElse("false").toBoolean()
     if (!ideaActive) {
       api(project(":polaris-spark-3.5_2.13"))
-      api(project(":polaris-spark-integration-3.5_2.13"))
       api(project(":polaris-spark-4.0_2.13"))
-      api(project(":polaris-spark-integration-4.0_2.13"))
     }
 
     api(project(":polaris-admin"))
