+++ plugins/spark/v4.0/spark/build.gradle.kts
@@ -18,8 +18,13 @@
  */
 
 import com.github.jengelman.gradle.plugins.shadow.tasks.ShadowJar
+import org.gradle.api.attributes.java.TargetJvmVersion
+import org.gradle.api.plugins.jvm.JvmTestSuite
 
-plugins { id("polaris-client") }
+plugins {
+  id("polaris-client")
+  id("polaris-server-test-runner")
+}
 
 checkstyle {
   configProperties =
@@ -29,11 +34,6 @@ checkstyle {
     )
 }
 
-sourceSets {
-  main { java { srcDir("../../common/src/main/java") } }
-  test { java { srcDir("../../common/src/test/java") } }
-}
-
 // get version information
 val sparkMajorVersion = "4.0"
 val scalaVersion = getAndUseScalaVersionForProject()
@@ -41,8 +41,12 @@ val icebergVersion = libs.versions.iceberg.get()
 val spark40Version = libs.versions.spark40.get()
 
 val scalaLibraryVersion = libs.versions.scala213.get()
+val errorProneAnnotationsVersion = libs.errorprone.get().versionConstraint.requiredVersion
+val intTestJvmVersion = 21
 
 dependencies {
+  polarisServer(project(path = ":polaris-server", configuration = "quarkusRunner"))
+
   // TODO: extract a polaris-rest module as a thin layer for
   //  client to depends on.
   implementation(project(":polaris-core")) { isTransitive = false }
@@ -83,6 +87,127 @@ dependencies {
     exclude("org.slf4j", "jul-to-slf4j")
   }
   testImplementation("org.apache.logging.log4j:log4j-api:2.26.0")
+}
+
+testing {
+  suites {
+    register<JvmTestSuite>("intTest") {
+      dependencies {
+        implementation(
+          "org.apache.iceberg:iceberg-spark-runtime-${sparkMajorVersion}_${scalaVersion}:${icebergVersion}"
+        )
+
+        implementation(project(":polaris-api-management-model"))
+
+        implementation(project(":polaris-runtime-test-common"))
+
+        implementation("org.apache.spark:spark-sql_${scalaVersion}:${spark40Version}")
+
+        // Add spark-hive for Hudi integration - provides HiveExternalCatalog that Hudi needs
+        runtimeOnly("org.apache.spark:spark-hive_${scalaVersion}:${spark40Version}")
+        // Delta and Hudi initialize Log4j classes from the Spark runtime. Keep Log4j Core aligned
+        // with the version used by the Spark tests instead of relying on older Hive transitive
+        // deps.
+        runtimeOnly("org.apache.logging.log4j:log4j-core:2.26.0")
+
+        implementation("io.delta:delta-spark_${sparkMajorVersion}_${scalaVersion}:4.2.0")
+        implementation("org.apache.hudi:hudi-spark4.0-bundle_${scalaVersion}:1.1.1")
+
+        // The hudi-spark-bundle includes most Hive libraries but excludes hive-exec to keep size
+        // manageable
+        // This matches what Spark 4.0 distribution provides (hive-exec-2.3.10-core.jar)
+        implementation("org.apache.hive:hive-exec:2.3.10:core")
+
+        implementation(platform(libs.jackson.bom))
+        implementation("com.fasterxml.jackson.jakarta.rs:jackson-jakarta-rs-json-provider")
+        implementation(libs.jakarta.ws.rs.api)
+        compileOnly("com.google.errorprone:error_prone_annotations:${errorProneAnnotationsVersion}")
+
+        implementation(testFixtures(project(":polaris-runtime-service")))
+
+        implementation(platform(libs.awssdk.bom))
+        implementation("software.amazon.awssdk:glue")
+        implementation("software.amazon.awssdk:kms")
+        implementation("software.amazon.awssdk:dynamodb")
+
+        implementation(platform(libs.testcontainers.bom))
+        implementation("org.testcontainers:testcontainers")
+        implementation(libs.s3mock.testcontainers)
+
+        // Required for Spark integration tests
+        implementation(enforcedPlatform("org.scala-lang:scala-library:${scalaLibraryVersion}"))
+        implementation(enforcedPlatform("org.scala-lang:scala-reflect:${scalaLibraryVersion}"))
+        implementation(libs.javax.servlet.api)
+        implementation(libs.antlr4.runtime.spark40)
+      }
+
+      targets.all {
+        testTask.configure {
+          environment(
+            "AWS_REGION",
+            providers.environmentVariable("AWS_REGION").getOrElse("us-west-2"),
+          )
+          jvmArgs("--add-exports", "java.base/sun.nio.ch=ALL-UNNAMED")
+          // Need to allow a java security manager after Java 21, for Subject.getSubject to work
+          // "getSubject is supported only if a security manager is allowed".
+          systemProperty("java.security.manager", "allow")
+          val logsDir = project.layout.buildDirectory.get().asFile.resolve("logs")
+          // delete files from previous runs
+          doFirst {
+            // delete log files written by Polaris
+            logsDir.deleteRecursively()
+          }
+          withPolarisServer(configurations.polarisServer) {
+            environment.put(
+              "AWS_REGION",
+              providers.environmentVariable("AWS_REGION").orElse("us-west-2"),
+            )
+            environment.put(
+              "AWS_ACCESS_KEY_ID",
+              providers.environmentVariable("AWS_ACCESS_KEY_ID").orElse("ap1"),
+            )
+            environment.put(
+              "AWS_SECRET_ACCESS_KEY",
+              providers.environmentVariable("AWS_SECRET_ACCESS_KEY").orElse("s3cr3t"),
+            )
+            environment.put("AWS_EC2_METADATA_DISABLED", "true")
+            environment.put("POLARIS_BOOTSTRAP_CREDENTIALS", "POLARIS,test-admin,test-secret")
+            systemProperties.put("quarkus.profile", "it")
+            systemProperties.put(
+              "quarkus.log.file.path",
+              logsDir.resolve("polaris.log").absolutePath,
+            )
+          }
+          // For Spark integration tests
+          addSparkJvmOptions()
+        }
+      }
+    }
+  }
+}
+
+sourceSets {
+  main { java { srcDir("../../common/src/main/java") } }
+  test { java { srcDir("../../common/src/test/java") } }
+  named("intTest") {
+    java { srcDir("../../common/src/intTest/java") }
+    resources { srcDir("../../common/src/intTest/resources") }
+  }
+}
+
+listOf("intTestCompileClasspath", "intTestRuntimeClasspath").forEach {
+  // :polaris-runtime-test-common and :polaris-runtime-service (testFixtures) require JVM 21
+  // compatibility.
+  configurations.named(it).configure {
+    attributes.attribute(TargetJvmVersion.TARGET_JVM_VERSION_ATTRIBUTE, intTestJvmVersion)
+  }
+}
+
+// Force jakarta.servlet-api to 5.0.0 for Spark 4.0 compatibility
+// Spark 4.0 requires version 5.0.0 which includes SingleThreadModel
+// Quarkus BOM forces it to 6.x which removed SingleThreadModel
+configurations.named("intTestRuntimeClasspath") {
+  resolutionStrategy { force("jakarta.servlet:jakarta.servlet-api:5.0.0") }
 }
 
 tasks.register<ShadowJar>("createPolarisSparkJar") {
